+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added `PointcloudObject`.
 * Added `compas_viewer.base.Base` to access viewer singleton.
 * Added `compas_viewer.singleton.SingletonMeta` and `compas_viewer.singleton.Singleton`.
+* Added `compas_viewer.scene.Group` and `compas_viewer.scene.GroupObject`.
 
 ### Changed